refactor: 테스트 코드의 moverules 패키지명 camelCase에서 전부 소문자로 변경
