+++ src/test/java/janggi/domain/moverules/PoMoveRuleTest.java
@@ -1,23 +1,23 @@
-package janggi.domain.moveRules;
+package janggi.domain.moverules;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.common.Direction;
 import janggi.domain.common.Team;
-import janggi.domain.piece.moverules.ChaMoveRule;
 import janggi.domain.piece.moverules.MoveRule;
+import janggi.domain.piece.moverules.PoMoveRule;
 import janggi.domain.route.Route;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ChaMoveRuleTest {
+public class PoMoveRuleTest {
 
     @Test
-    @DisplayName("차는 직선으로 갈 수 있다")
-    void 차의_이동규칙() {
+    @DisplayName("포는 직선으로 갈 수 있다")
+    void 포의_이동규칙() {
         //given
-        MoveRule moveRule = new ChaMoveRule();
+        MoveRule moveRule = new PoMoveRule();
         Team team = Team.CHO;
         Route route1 = new Route(List.of(Direction.UP));
         Route route2 = new Route(List.of(Direction.RIGHT));
@@ -26,9 +26,9 @@ class ChaMoveRuleTest {
         List<Route> routes = List.of(route1, route2, route3, route4);
 
         //when
-        List<Route> chaRoutes = moveRule.findRoutes(team);
+        List<Route> poRoutes = moveRule.findRoutes(team);
 
         //then
-        assertThat(chaRoutes).isEqualTo(routes);
+        assertThat(poRoutes).isEqualTo(routes);
     }
 }
