+++ src/test/java/janggi/domain/moverules/SaMoveRuleTest.java
@@ -1,23 +1,23 @@
-package janggi.domain.moveRules;
+package janggi.domain.moverules;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.common.Direction;
 import janggi.domain.common.Team;
-import janggi.domain.piece.moverules.KingMoveRule;
 import janggi.domain.piece.moverules.MoveRule;
+import janggi.domain.piece.moverules.SaMoveRule;
 import janggi.domain.route.Route;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class KingMoveRuleTest {
+public class SaMoveRuleTest {
 
     @Test
-    @DisplayName("왕은 직선과 대각선으로 갈 수 있다.")
-    void 왕의_이동규칙() {
+    @DisplayName("사는 직선과 대각선으로 갈 수 있다")
+    void 사의_이동규칙() {
         //given
-        MoveRule moveRule = new KingMoveRule();
+        MoveRule moveRule = new SaMoveRule();
         Team team = Team.CHO;
         Route route1 = new Route(List.of(Direction.UP));
         Route route2 = new Route(List.of(Direction.RIGHT));
@@ -30,9 +30,9 @@ public class KingMoveRuleTest {
         List<Route> routes = List.of(route1, route2, route3, route4, route5, route6, route7, route8);
 
         //when
-        List<Route> kingRoutes = moveRule.findRoutes(team);
+        List<Route> saRoutes = moveRule.findRoutes(team);
 
         //then
-        assertThat(kingRoutes).isEqualTo(routes);
+        assertThat(saRoutes).isEqualTo(routes);
     }
 }
